Made a track
I made a track for the notes to scroll down as well as a box at the bottom for where you are supposed to hit the notes.
I have almost finished adding the white line to the background too, I just  need to figure out how to blur it still.