added hit sound and key binds
I added a few lines of code that play a hit sound when you hit the notes on time. I also added it so that the players can see what the key binds are to properly hit the notes.  I also am thinking of adding a multiplier for when a player has a combo.