Getting a start on making a song
I changed to frame rate to match the bpm of the song. I spent quite a bit of time figuring out different functions in the p5 sound library reference page trying to figure out a good way of displaying notes in time with the song. I figured that I might use the addCue() function so that it triggers an event at a certain time.